test(BE): #S08P31A306-152 add project update service test

+++ back-end/roughcode/src/main/java/com/cody/roughcode/project/service/S3FileServiceImpl.java
@@ -1,8 +1,10 @@
 package com.cody.roughcode.project.service;
 
 import com.amazonaws.SdkClientException;
+import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.AmazonS3Client;
 import com.amazonaws.services.s3.model.*;
+import com.cody.roughcode.exception.DeletionFailException;
 import com.cody.roughcode.project.dto.req.ProjectReq;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -13,6 +15,8 @@ import org.springframework.web.multipart.MultipartFile;
 import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
+import java.net.MalformedURLException;
+import java.net.URL;
 import java.net.URLDecoder;
 import java.nio.file.Path;
 import java.time.LocalDateTime;
@@ -36,7 +40,7 @@ public class S3FileServiceImpl implements S3FileService {
 
     // 이미지 업로드 후 URL 리턴
     @Override
-    public String upload(MultipartFile multipartFile, String dirName, List<String> fileNames) throws IOException {
+    public String upload(MultipartFile multipartFile, String dirName, String fileName) throws IOException {
         log.info("-----------upload method start-----------");
         log.info("file : {}, dirName : {}", multipartFile, dirName);
 
@@ -45,21 +49,20 @@ public class S3FileServiceImpl implements S3FileService {
                 .orElseThrow(() -> new IllegalArgumentException("MultipartFile에서 File로 변환에 실패했습니다."));
 
         // 파일명에 project 정보 같이 입력
-        StringBuilder fileName = new StringBuilder(dirName + "/project");
-        for(String str : fileNames){
-            fileName.append(str);
-            fileName.append("_");
-        }
-        fileName.deleteCharAt(fileName.length() - 1);
+        StringBuilder fileInfo = new StringBuilder(dirName + "/" + fileName);
 
-        log.info("new file Name : {}", fileName);
+        log.info("new file Name : {}", fileInfo);
 
         // S3로 업로드
-        amazonS3Client.putObject(new PutObjectRequest(bucket, String.valueOf(fileName), uploadFile)
+        amazonS3Client.putObject(new PutObjectRequest(bucket, String.valueOf(fileInfo), uploadFile)
                 .withCannedAcl(CannedAccessControlList.PublicRead));
-        
-        String uploadImageUrl = amazonS3Client.getUrl(bucket, String.valueOf(fileName)).toString();
-        
+
+        URL imageUrl = amazonS3Client.getUrl(bucket, String.valueOf(fileInfo));
+        if (imageUrl == null) {
+            throw new NullPointerException("이미지 저장에 실패했습니다.");
+        }
+        String imageUrlString = imageUrl.toString();
+
         // 로컬 파일 삭제
         if (uploadFile.exists()) {
             if (uploadFile.delete()) {
@@ -69,7 +72,9 @@ public class S3FileServiceImpl implements S3FileService {
             }
         }
 
-        return uploadImageUrl;
+        log.info("return : {}", imageUrlString);
+
+        return imageUrlString;
     }
 
     // multipartFile -> File 형식으로 변환 및 로컬에 저장
@@ -83,32 +88,17 @@ public class S3FileServiceImpl implements S3FileService {
     }
 
     // 이미지 삭제 method
-    @Override
-    public boolean delete(String profileUrl, String dirName) {
-        log.info("profile url : {}", profileUrl);
-        
-        // S3에서 이미지 검색
-        Pattern tokenPattern = Pattern.compile("(?<=profile/).*");
-        Matcher matcher = tokenPattern.matcher(profileUrl);
-
-        String foundImage = null;
-        if (matcher.find()) {
-            foundImage = matcher.group();
-        }
-
-        String originalName = URLDecoder.decode(foundImage);
-        String filePath = dirName + "/" + originalName;
-        log.info("originalName : {}", originalName);
-        
-        // S3에서 이미지 삭제
+    public boolean delete(String imageUrlString) {
         try {
-            amazonS3Client.deleteObject(new DeleteObjectRequest(bucket, filePath));
-            log.info("deletion complete : {}", filePath);
-            return true;
-        } catch (SdkClientException e) {
+            URL imageUrl = new URL(imageUrlString);
+            String key = imageUrl.getPath().substring(1);
+            amazonS3Client.deleteObject(bucket, key);
+        } catch (Exception e) {
             log.error(e.getMessage());
-            return false;
+            throw new DeletionFailException("이미지");
         }
+
+        return true;
     }
 
 }
